+++ src/pages/user/UserPage.module.css
@@ -3,7 +3,6 @@
 }
 
 .title {
-
     display: flex;
     margin: 20px;
     justify-content: center;
@@ -36,6 +35,7 @@
 }
 
 .user_title {
+  display: flex;
   width: 300px;
 }
 
@@ -45,11 +45,12 @@
 }
 .postmenu li {
     display: flex;
-    align-items: center;
     cursor: pointer;
-    margin: 15px;
+    margin-left: 40px;
+    margin-right: 40px;
 }
 .postmenu ul {
+    display: flex;
     justify-content: center;
 }
 .button {
